+++ src/Trigger.ts
@@ -63,7 +63,7 @@ export default class Trigger {
     log.verbose(`Trigger ${this.name}`, `${fileName}: Analyzing`);
     const startTime = new Date();
     const analysis = await analyzeImage(fileName).catch(e => {
-      log.warn("Trigger ${this.name}", e);
+      log.warn(`Trigger ${this.name}`, e);
       return undefined;
     });
 
@@ -308,11 +308,13 @@ export default class Trigger {
    * Stops watching for file changes.
    */
   public async stopWatching(): Promise<void> {
-    await this._watcher.close().catch(e => {
-      throw Error(`Trigger ${this.name} unable to stop watching for images: ${e}`);
-    });
+    if (this._watcher) {
+      await this._watcher.close().catch(e => {
+        log.warn(`Trigger ${this.name}`, `unable to stop watching for images: ${e}`);
+      });
 
-    log.verbose(`Trigger ${this.name}`, `Stopped listening for new images in ${this.watchPattern}`);
+      log.verbose(`Trigger ${this.name}`, `Stopped listening for new images in ${this.watchPattern}`);
+    }
   }
 
   /**