+++ src/WebServer.ts
@@ -9,8 +9,10 @@ import * as Settings from "./Settings";
 import express from "express";
 import motionRouter from "./routes/motion";
 import path from "path";
+import { Server } from "http";
 
 const app = express();
+let server: Server;
 
 export function startApp(): void {
   const annotatedImagePath = path.join(LocalStorageManager.localStoragePath, LocalStorageManager.Locations.Annotations);
@@ -19,8 +21,15 @@ export function startApp(): void {
   app.use("/motion", motionRouter);
 
   try {
-    app.listen(Settings.port, () => log.info("Web server", `Listening at http://localhost:${Settings.port}`));
+    server = app.listen(Settings.port, () => log.info("Web server", `Listening at http://localhost:${Settings.port}`));
   } catch (e) {
     log.warn("Web server", `Unable to start web server: ${e.error}`);
+  }
+}
+
+export function stopApp(): void {
+  if (server) {
+    log.verbose("Web server", "Stopping.");
+    server.close();
   }
 }