+++ src/WebServer.ts
@@ -7,13 +7,19 @@ import * as log from "./Log";
 import * as Settings from "./Settings";
 
 import express from "express";
+import motionRouter from "./routes/motion";
+import path from "path";
 import { Server } from "http";
 
 const app = express();
 let server: Server;
 
 export function startApp(): void {
-  app.use("/", express.static(LocalStorageManager.localStoragePath));
+  const annotatedImagePath = path.join(LocalStorageManager.localStoragePath, LocalStorageManager.Locations.Annotations);
+  app.use("/", express.static(annotatedImagePath));
+  app.use("/annotations", express.static(annotatedImagePath));
+  app.use("/motion", motionRouter);
+
   try {
     server = app.listen(Settings.port, () => log.info("Web server", `Listening at http://localhost:${Settings.port}`));
   } catch (e) {