+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Shutdown is handled gracefully now, with file system watchers and web server stopped
+  when various signals are received (e.g. ctrl+C). Resolves [issue 280](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/265)
 - Changes to `settings.json` and `triggers.json` are automatically detected and cause a reload. No need to restart
   the Docker container anymore! Just save the file and the system should notice the change and reload with the new
   configuration. Resolves [issue 278](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/278).