+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Requests to end the system (e.g. via ctrl+c) are handled gracefully now. Resolves [issue 280](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/280).
 - Changes to `settings.json` and `triggers.json` are automatically detected and cause a reload. No need to restart
   the Docker container anymore! Just save the file and the system should notice the change and reload with the new
   configuration. Resolves [issue 278](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/278).